+++ src/org/openscience/cdk/test/reaction/type/ElectronImpactPDBReactionTest.java
@@ -88,11 +88,11 @@ public class ElectronImpactPDBReactionTest extends CDKTestCase {
         Assert.assertEquals(2, setOfReactions.getReactionCount());
 
         IMolecule molecule = setOfReactions.getReaction(0).getProducts().getMolecule(0);
-        Assert.assertEquals(1, molecule.getAtom(1).getFormalCharge());
+        Assert.assertEquals(1, molecule.getAtom(1).getFormalCharge().intValue());
         Assert.assertEquals(1, molecule.getConnectedSingleElectronsCount(molecule.getAtom(0)));
 
         molecule = setOfReactions.getReaction(1).getProducts().getMolecule(0);
-        Assert.assertEquals(1, molecule.getAtom(0).getFormalCharge());
+        Assert.assertEquals(1, molecule.getAtom(0).getFormalCharge().intValue());
         Assert.assertEquals(1, molecule.getConnectedSingleElectronsCount(molecule.getAtom(1)));
         
         Assert.assertEquals(3,setOfReactions.getReaction(0).getMappingCount());
@@ -120,11 +120,11 @@ public class ElectronImpactPDBReactionTest extends CDKTestCase {
 
         
         IMolecule molecule = setOfReactions.getReaction(0).getProducts().getMolecule(0);
-        Assert.assertEquals(1, molecule.getAtom(1).getFormalCharge());
+        Assert.assertEquals(1, molecule.getAtom(1).getFormalCharge().intValue());
         Assert.assertEquals(1, molecule.getConnectedSingleElectronsCount(molecule.getAtom(0)));
 
         molecule = setOfReactions.getReaction(1).getProducts().getMolecule(0);
-        Assert.assertEquals(1, molecule.getAtom(0).getFormalCharge());
+        Assert.assertEquals(1, molecule.getAtom(0).getFormalCharge().intValue());
         Assert.assertEquals(1, molecule.getConnectedSingleElectronsCount(molecule.getAtom(1)));
 		
 	}
@@ -149,11 +149,11 @@ public class ElectronImpactPDBReactionTest extends CDKTestCase {
         Assert.assertEquals(2, setOfReactions.getReactionCount());
 
         IMolecule molecule = setOfReactions.getReaction(0).getProducts().getMolecule(0);
-        Assert.assertEquals(1, molecule.getAtom(1).getFormalCharge());
+        Assert.assertEquals(1, molecule.getAtom(1).getFormalCharge().intValue());
         Assert.assertEquals(1, molecule.getConnectedSingleElectronsCount(molecule.getAtom(0)));
 
         molecule = setOfReactions.getReaction(1).getProducts().getMolecule(0);
-        Assert.assertEquals(1, molecule.getAtom(0).getFormalCharge());
+        Assert.assertEquals(1, molecule.getAtom(0).getFormalCharge().intValue());
         Assert.assertEquals(1, molecule.getConnectedSingleElectronsCount(molecule.getAtom(1)));
         
         Assert.assertEquals(3,setOfReactions.getReaction(0).getMappingCount());
